+++ src/EventSubscriber/CheckoutConfirmEventSubscriber.php
@@ -58,6 +58,10 @@ class CheckoutConfirmEventSubscriber implements EventSubscriberInterface
         if ($selectedPaymentGateway->getHandlerIdentifier() === CreditCard::class) {
             $this->addPaymentMethodExtension($templateVariables, $pageObject, 'creditCard', $amount, $threeDS, $isGuest, $shippingError, $context, $salesChannelContext);
         }
+
+        if ($selectedPaymentGateway->getHandlerIdentifier() === AchEcheck::class) {
+            $this->addPaymentMethodExtension($templateVariables, $pageObject, 'achEcheck', $amount, $threeDS, $isGuest, $shippingError, $context, $salesChannelContext);
+        }
     }
 
     private function addPaymentMethodExtension(
@@ -109,6 +113,12 @@ class CheckoutConfirmEventSubscriber implements EventSubscriberInterface
             ]);
         }
 
+        if ($gateway === 'achEcheck') {
+            $templateVariables->assign([
+                'configs' => $this->configService->getModeConfig(),
+            ]);
+        }
+
         $pageObject->addExtension(CheckoutTemplateCustomData::EXTENSION_NAME, $templateVariables);
     }
 
@@ -117,6 +127,8 @@ class CheckoutConfirmEventSubscriber implements EventSubscriberInterface
         switch ($gateway) {
             case 'creditCard':
                 return '@Storefront/nmi-payment/credit-card.html.twig';
+            case 'achEcheck':
+                return '@Storefront/nmi-payment/ach-eCheck.html.twig';
             default:
                 return '';
         }