+++ src/Service/NMIACHPaymentDataRequestService.php
@@ -4,36 +4,30 @@ namespace NMIPayment\Service;
 
 use Psr\Log\LoggerInterface;
 use Shopware\Core\Checkout\Cart\Cart;
-use Shopware\Core\Framework\Uuid\Uuid;
 use Shopware\Core\System\NumberRange\ValueGenerator\NumberRangeValueGeneratorInterface;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 
-class NMIPaymentDataRequestService
+class NMIACHPaymentDataRequestService
 {
   private NMIConfigService $nmiConfigService;
   private NMIPaymentApiClient $nmiPaymentApiClient;
-  private VaultedCustomerService $vaultedCustomerService;
   private NumberRangeValueGeneratorInterface $numberRangeGenerator;
   private LoggerInterface $logger;
 
   public function __construct(
     NMIConfigService $nmiConfigService,
     NmiPaymentApiClient $nmiPaymentApiClient,
-    VaultedCustomerService $vaultedCustomerService,
     NumberRangeValueGeneratorInterface $numberRangeGenerator,
     LoggerInterface $logger)
-
   {
     $this->nmiConfigService = $nmiConfigService;
     $this->nmiPaymentApiClient = $nmiPaymentApiClient;
-    $this->vaultedCustomerService = $vaultedCustomerService;
     $this->numberRangeGenerator = $numberRangeGenerator;
     $this->logger = $logger;
   }
 
-  public function sendPaymentRequestToNMI(array $data, Cart $cart, SalesChannelContext $context): array
+  public function sendACHPaymentRequestToNMI(array $data, Cart $cart, SalesChannelContext $context): array
   {
-
     $salesChannelId = $context->getSalesChannel()->getId();
     $this->nmiPaymentApiClient->initializeForSalesChannel($salesChannelId);
     $mode = $this->nmiConfigService->getConfig('mode', $salesChannelId);
@@ -44,8 +38,6 @@ class NMIPaymentDataRequestService
     );
 
     $customer = $context->getCustomer();
-
-
     if (!$customer) {
       return ['success' => false, 'message' => 'Customer not found'];
     }
@@ -56,13 +48,9 @@ class NMIPaymentDataRequestService
     $paymentMethodType = $this->nmiConfigService->getConfig('authorizeAndCapture');
     $allLineItemsCart = $cart->getLineItems();
     $nonSubscriptionItems = [];
-    $isSubscriptionCart = false;
 
     foreach ($allLineItemsCart as $lineItem) {
       $payload = $lineItem->getPayload();
-      if (!empty($payload['isSubscription'])) {
-        $isSubscriptionCart = true;
-      }
       $productData = [
         'productNumber' => $payload['productNumber'] ?? null,
         'description' => $lineItem->getLabel(),
@@ -72,11 +60,6 @@ class NMIPaymentDataRequestService
       ];
       $nonSubscriptionItems[] = $productData;
     }
-
-    $responses = [];
-    $customerVaultId = $this->vaultedCustomerService->getVaultedCustomerIdByCustomerId($context->getContext(), $customer->getId())
-      ? $this->vaultedCustomerService->getVaultedCustomerIdByCustomerId($context->getContext(), $customer->getId())
-      : null;
 
     $predictedOrderNumber = $this->numberRangeGenerator->getValue(
       'order',
@@ -91,6 +74,13 @@ class NMIPaymentDataRequestService
       'currency' => 'USD',
       'type' => $paymentMethodType ? 'auth' : 'sale',
       'orderid' => $predictedOrderNumber,
+      'payment' => 'check',
+      'payment_token' => $data['token'] ?? null,
+      'checkname' => $data['checkname'] ?? '',
+      'checkaba' => $data['checkaba'] ?? '',
+      'checkaccount' => $data['checkaccount'] ?? '',
+      'account_holder_type' => $data['account_holder_type'] ?? 'personal',
+      'account_type' => $data['account_type'] ?? 'checking',
       'first_name' => $data['first_name'],
       'last_name' => $data['last_name'],
       'address1' => $billingAddress ? $billingAddress->getStreet() : '',
@@ -108,39 +98,13 @@ class NMIPaymentDataRequestService
       'shipping_state' => $shippingAddress->getCountryState() ? $shippingAddress->getCountryState()->getShortCode() : '',
       'shipping_country' => $shippingAddress->getCountry() ? $shippingAddress->getCountry()->getIso() : '',
       'shipping_email' => $customer->getEmail(),
-      'payment_token' => $data['token'] ?? null,
-      'acu_enabled' => 'true',
       'line_items' => $nonSubscriptionItems,
     ];
 
-    $billingId = null;
-    $shouldVault = $data['saveCard'] || $isSubscriptionCart;
-
-    if ($shouldVault) {
-      $billingId = Uuid::randomHex();
-
-      if ($customerVaultId === null) {
-        $postData['customer_vault'] = 'add_customer';
-        $postData['billing_id'] = $billingId;
-      } else {
-        $postData['customer_vault_id'] = $customerVaultId;
-        $postData['customer_vault'] = 'add_billing';
-        $postData['billing_id'] = $billingId;
-      }
-    }
-
     $response = $this->nmiPaymentApiClient->createTransaction($postData);
-    $this->logger->info('Payment Response -> ' . json_encode($response));
+    $this->logger->info('ACH Payment Response -> ' . json_encode($response));
 
     $processedResponse = $this->handleNMIResponse($response);
-
-    if (!empty($processedResponse['customer_vault_id'])) {
-      $customerVaultId = $processedResponse['customer_vault_id'];
-      $processedResponse['billing_id'] = $billingId;
-      $processedResponse['isSubscriptionCart'] = $isSubscriptionCart;
-    }
-
-    $responses['payment'] = $processedResponse;
 
     if (!$processedResponse['success']) {
       return $processedResponse;
@@ -149,9 +113,7 @@ class NMIPaymentDataRequestService
     return [
       'success' => true,
       'message' => 'Transaction processed successfully',
-      'customer_vault_id' => $data['saveCard']?  $customerVaultId : null,
-      'billing_id' => $billingId ?  $billingId : null,
-      'responses' => $responses
+      'transaction_id' => $processedResponse['transaction_id'],
     ];
   }
 
@@ -162,17 +124,13 @@ class NMIPaymentDataRequestService
         'success' => true,
         'message' => 'Payment successful!',
         'transaction_id' => $response['transactionid'],
-        'customer_vault_id' => $response['customer_vault_id'] ?? null,
-        'billing_id' => $response['billing_id'] ?? null,
-        'isSubscriptionCart' => $response['isSubscriptionCart'] ?? false,
       ];
     } else {
-//            $this->logger->warning('Payment failed', ['response' => $response]);
       return [
         'success' => false,
         'message' => 'Payment failed: ' . ($response['responsetext'] ?? 'Unknown error')
       ];
     }
   }
-
 }
+