+++ CHANGELOG.md
@@ -1,5 +1,21 @@
-Changelog
+# Changelog
+
 All notable changes to this project will be documented in this file.
+
+## [1.2.2] - 2025-12-19
+
+### Added Features
+- Webhooks and validation to handle:
+    - Void
+    - Refund
+    - Partial refund
+- New payment method added - **ACH**
+- Automatic webhook registration when the **Test Connection** button returns a successful response
+
+### Fixed
+- Resolved an issue where production and sandbox API keys did not correctly set the environment mode
+
+---
 
 [1.0.0] - 03-11-2025
 Added